otimizações no widget da animação
